Refactor Address class to simplify the treatment of "token-aware" addrs
We switched from the Address class "capturing" the token-aware quality
to it just being a display encoding format.  The philosophy of this
class is that it encapsulates a locking script (akin to CTxTestination
in BCHN).  As such, the way the address is encoded as a string doesn't
need to be captured by this class.

This simplifies the usage of the class itself and avoids potential bugs
where the same locking script is not "seen" by electron cash as being
the same address if one was token and the other was cashaddr, etc.

This refactor makes actually *using* the address class easier than what
we had before where token-awareness was a "kind".  That was confusing to
callers.  It really should just be an encoding type...

Also, the tests were updated and the Address class was made to enforce its
invariants much more strictly.