wallet "sweep" -- try to explicitly exclude cashtokens
This is a defensive-programming commit. We declare our protocol version
as 1.4 to the server, so `blockchain.scripthash.listunspent` should play
nice with us and never show us tokens. However, just to be safe, we
explicitly ask to "exclude_tokens" and if that call fails because the
server doesn't understand the argument, we try again without that
parameter.

This is defensive programming in case someone someday upgrades our
Electron Cash protocol version to 1.5 (in which case tokens are included
in the unspent list by default!)