Fixups to token history tab; added fungible total balance and nft balance
Makes it more like the main history tab and also it's useful to know
your on-going balance.

Leverages the new wallet API that quickly determines these things.

Also some fixes to the column widths and fonts... to fit more on the
screen horizontally. Also added more tooltips and also made more of the
columns resizeable.