Added token metadata management mechanism, added default token icons
Icons get saved to DATADIR/cashtoken_meta/icons as the "category-id".png.
However, if no icon is found for a particular category-id there, electron
cash just uses some defaults which are procedurally generated based on the
category-id.  The icons produced deterministically in this way are identical
to the code that is being used at https://my-cashtokens-webwallet.netlify.app/