Added an optional "URL" field to the Create Token form
If specified, we end up hitting the web to calculate the document's
hash, and we embed the hash and url into a BCMR-conforming OP_RETURN
field.

Also in this commit: Added support for the tx_dialog broadcast callback
to accept betwee 1-3 args (just in case we need to inspect what was
broadcast in a later form).