Load and save txns to wallet in sorted order
This ensures that the ct_txid_hash we store in the wallet to detect
opening of wallet with non-CashToken electron cash is always accurate,
since txid's can get shuffled around in the wallet.transactions dict.

The performance penalty for sorting is on the order of 2 msec per 10k
txns in wallet.  It's not too bad and this fixes the issue of the wallet
sometimes erroneously thinking it needs to re-calculate ct_txi and
ct_txo.