Add support for Cash token in Satochip
see:
https://github.com/bitjson/cashtokens#token-encoding

For cash token tx preimage signature, we use the pysatochip 'sign_transaction_hash' method since the tx preimage format is not backward compatible with legacy Bitcoin Cash preimage tx serialization.

Note that if the 2FA (2-Factor-Authentication) option is enabled in the Satochip, cash-token tx is not fully supported yet by the 2FA android app. The app will not be able to parse the new tx format (show an error instead) but will allow blind signing (signing the request without displaying it).
This will be corrected in a future release.

For maximum backward compatibility with 2FA, we only use 'sign_transaction_hash' for preimage tx with cash token. For non-cash-token preimage, we still use the (well supported) 'sign_transaction' method.