build: Update to Python 3.11.8
There seems to be some issue with Wine but it works:

```
💬 INFO:  Installing core...
💬 INFO:  Installing dev...
wine: Unhandled stack overflow at address 70BBE164 (thread 0118), starting debugger...
💬 INFO:  Installing exe...
💬 INFO:  Installing lib...
💬 INFO:  Installing pip...
💬 INFO:  Installing tools...
```

This updates pip to version 23.2.1, which is needed for building with
newer Python versions.

This also removes the pip patch for reproducible .zip files, which is
not needed anymore starting with pip version 22.2. Version 22.2 updated
the vendored `distlib` to version 0.3.5, which has a fix for the issue.

see:
https://github.com/Electron-Cash/Electron-Cash/pull/2424
https://github.com/pypa/pip/pull/11267/commits/fd0ea6bc5e8cb95e518c23d901c26ca14db17f89
https://github.com/pypa/distlib/issues/164
https://github.com/pypa/distlib/pull/165

This also updates the ecdsa, cryptography and setuptools package.