Added UI hooks so that HW wallets that can't sign cashtokens warn user
This addresses the issue where HW wallet users assume they can receive
CashTokens, because the UI shows them token-aware addresses as well as a
create token form.

However, most HW wallets except for Satochip cannot *send* tokens
because they lack the ability to sign token-bearing inputs.

As such, it is preferable to warn the user when a HW wallet is
attempting to create a token or receive a token, that their wallet
cannot sign for them.

This avoids users having tokens locked-in to a HW wallet, unable to send
them further, which is bad UX.