Implemented a new wallet type: `multi_xpub` (#2731)
Squashed  commit of the following:

* wip

* Made the base_wizard simpler for multi-xpub

Bury the fact that we support multiple xpubs deep in the wizard so as to
not annoy users, and keep this as a power option.

* tweak: allow pw protect on multi xpub that has xprvs

* Tweaks to CashFusion to get it to recognize and use MultiXPubWallet

Also changed the error message CashFusion displays when it can't use a
wallet since it was sometimes confusing -- e.g. hw wallets are
"standard" but it claimed CashFusion doesn't work with "standard"
wallets, which is not true.

* More fixups and more progress

* Fixup to exporting WIF keys

* Added MultiXPubWallet.add_keystore and .remove_keystore, also a small fix

The small fix is to implement pruned_txo removal in
ImportedWalletBase.delete_address, which was a FIXME that was left
unfixed.

Also delete ct_txi and ct_txo in said function as well.

* Implemented UI to add/delete keys from MultiXPubWallet

Appears in Wallet -> Information

* Added an assert to MultiXPubKey wallet to enforce only HD keystores

* Some refactoring...

Refactored class Xprv out of BIP32_KeyStore.

This, I feel, makes the code more manageable (maybe).

* Implemented toggle xpub/xprv in Wallet -> Information dialog

* Refactor: Move the "Wallet Information" screen into a separate file

The method in main_window was getting a little long. Moved it to a
separate file to keep things sane(r).

* Nit: add base class get_master_public_keys method to Abstract_Wallet