Added UI menu actions to Update Category and/or NFT-specific BCMR data
Right-click in the CashTokens or Token History tab on a token or NFT to
see the new "Fetch Category Metadata" and/or "Fetch NFT Metadata" UI actions.

These actions download & apply the BCMR data as 1 UI action, no longer
requiring the user to go to the "Category Properties ..." menu to apply
BCMR data to a token.